backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
coverage==5.5
Django==3.1.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-vote==2.2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
isort==5.6.1
lazy-object-proxy==1.4.3
logilab-common==1.7.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.3.1
pydotplus==2.0.2
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-magic==0.4.22
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typing-extensions==3.7.4.3
wrapt==1.12.1
